foaf-vector-rag/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-docx>=1.0.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
python-dotenv>=1.2.0
httpx>=0.28.0
scikit-learn>=1.5.0
pytest>=8.0.0
